update server readme
